+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: (Zcash/Piratechain) Don't save sensitive initializer object to engine
+
 ## 4.26.0 (2024-10-03)
 
 - changed: Add redundancy to Cardano queries